format

+++ src/mvc/view/controls/SizeSelector.java
@@ -12,7 +12,6 @@ public class SizeSelector extends JSpinner {
 
     public SizeSelector(Controller controller) {
         this.controller = controller;
-
         configure();
     }
 
@@ -22,6 +21,6 @@ public class SizeSelector extends JSpinner {
     }
 
     private void changeListener() {
-        addChangeListener(e -> controller.notify(MAZE_SIDE_CHANGED, (Integer) getValue()));
+        addChangeListener(e -> controller.notify(MAZE_SIDE_CHANGED, getValue()));
     }
 }
